adds getBasses

+++ models/manufacturer.go
@@ -3,10 +3,9 @@ package models
 import "database/sql"
 
 type Manufacturer struct {
-	ID          uint64            `json:"id"`
+	ID          uint16         `json:"id"`
 	Name        string         `json:"name"`
 	FoundedYear int            `json:"founded_year"`
 	Nationality string         `json:"nationality"`
 	Logo        sql.NullString `json:"logo"`
 }
-
